pedro/requirements.txt
numpy
pandas

langchain
langchain-community
langchain-huggingface
langchain-core
sentence-transformers
transformers
faiss-cpu
accelerate
scikit-learn
tiktoken

nltk

pdfminer.six
PyPDF2
PyMuPDF
tabulate
pycryptodome

loguru
black[jupyter]
isort
ipython
ipykernel
python-dotenv
